Added givback liquidity part support
