+++ schema.graphql
@@ -29,6 +29,7 @@ type Balance @entity{
     rewardPerTokenPaidBalancer:  BigInt
     rewardsBalancer:  BigInt
     givback: BigInt!
+    givbackLiquidPart: BigInt!
     balancerLp: BigInt!
     balancerLpStaked: BigInt!
     sushiswapLp: BigInt!
