+++ src/mappings/tokenDistroMapping.ts
@@ -8,10 +8,13 @@ import {
   RoleGranted,
   RoleRevoked,
   StartTimeChanged,
+  TokenDistro,
 } from '../../generated/TokenDistro/TokenDistro';
-import { saveTokenAllocation, onGivBackPaid } from '../commons/tokenAllocation';
+import { saveTokenAllocation } from '../commons/tokenAllocation';
 import { addAllocatedTokens, addClaimed } from '../commons/balanceHandler';
 import { createTokenDistroContractInfoIfNotExists } from '../commons/TokenDistroHandler';
+import { Balance, TokenAllocation, TransactionTokenAllocation } from '../../generated/schema';
+import { GIVBACK } from '../helpers/constants';
 
 export function handleAllocate(event: Allocate): void {
   saveTokenAllocation(
@@ -35,7 +38,41 @@ export function handleClaim(event: Claim): void {
 }
 
 export function handleGivBackPaid(event: GivBackPaid): void {
-  onGivBackPaid(event.transaction.hash.toHex(), event.address);
+  // onGivBackPaid(event.transaction.hash.toHex(), event.address);
+
+  const transactionTokenAllocations = TransactionTokenAllocation.load(event.transaction.hash.toHex());
+
+  if (!transactionTokenAllocations) {
+    return;
+  }
+
+  const contract = TokenDistro.bind(event.address);
+  const globallyClaimableNow = contract.try_globallyClaimableAt(event.block.timestamp);
+  const totalTokens = contract.totalTokens();
+
+  for (let i = 0; i < transactionTokenAllocations.tokenAllocationIds.length; i++) {
+    const tokenAllocation = TokenAllocation.load(transactionTokenAllocations.tokenAllocationIds[i]);
+    if (!tokenAllocation) {
+      continue;
+    }
+    tokenAllocation.givback = true;
+    tokenAllocation.distributor = GIVBACK;
+    tokenAllocation.save();
+    const balance = Balance.load(tokenAllocation.recipient);
+    if (!balance) {
+      continue;
+    }
+
+    balance.givback = balance.givback.plus(tokenAllocation.amount);
+
+    if (!globallyClaimableNow.reverted) {
+      balance.givbackLiquidPart = balance.givbackLiquidPart.plus(
+        tokenAllocation.amount.times(globallyClaimableNow.value).div(totalTokens)
+      );
+    }
+
+    balance.save();
+  }
 }
 
 export function handleRoleAdminChanged(event: RoleAdminChanged): void {}
